% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{ais2segments}
\alias{ais2segments}
\title{Read ais.txt data from URL path and return spatial data.frame}
\usage{
ais2segments(data = NULL)
}
\arguments{
\item{data}{AIS dataframe from urls2df function}
}
\value{
sf data.frame with segments/linestrings with distance (km), calculated speed, and reported speed
}
\description{
Read ais.txt data from URL path and return spatial data.frame
}
\examples{
create segments from dataframe (data=df)
 segs=ais2segments(data = df)
Empty files returns empty df with date, filename, and timestamp and empty dataframes returns NULL

segs_empty = ais2segments(data = df_empty_url)

}
