% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_url_file_size}
\alias{check_url_file_size}
\title{Check file size from url function ---}
\usage{
check_url_file_size(path)
}
\arguments{
\item{path}{(url path)}
}
\value{
size of the file
}
\description{
Check file size from url function ---
}
\examples{
tst1 = check_url_file_size("https://ais.sbarc.org/logs_delimited/2019/191109/AIS_SBARC_191109-18.txt")
74251
tst2 = check_url_file_size("https://ais.sbarc.org/logs_delimited/2019/191130/AIS_SBARC_191130-00.txt")
0
}
