% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{create_segment}
\alias{create_segment}
\title{Create segment from 2 points}
\usage{
create_segment(p1, p2, crs = 4326)
}
\arguments{
\item{p1}{sf first point}

\item{p2}{sf second point}

\item{crs}{coordinate reference system (default 4326)}
}
\value{
linestring
}
\description{
Create segment from 2 points
}
