% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date.as_frac}
\alias{date.as_frac}
\title{Reformat the timestamp for fractionals ----}
\usage{
date.as_frac(t)
}
\arguments{
\item{t}{}
}
\value{
timestamp
}
\description{
Reformat the timestamp for fractionals ----
}
