% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date.build}
\alias{date.build}
\title{Build the datestring object ----}
\usage{
date.build(ymd = NULL, ts = NULL)
}
\arguments{
\item{ymd}{year,month,day}

\item{ts}{timestamp}
}
\description{
Build the datestring object ----
}
