% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to database}
\usage{
db_connect(db_yml = here::here("s4w_amazon_rds.yml"))
}
\arguments{
\item{yml}{yaml metadata for connecting. should contain parameters for: host, database, user, port, password}
}
\value{
con database connection object
}
\description{
Connect to database
}
