% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{.find_link_sibs}
\alias{.find_link_sibs}
\title{Find siblings, and build urls from a url}
\usage{
.find_link_sibs(start_url = NULL)
}
\arguments{
\item{start_url}{The URL the crawler begins it's call from, ie, our last URL from the logfile

.find_link_sibs("https://ais.sbarc.org/logs_delimited/2018/181212/")

> .find_link_sibs(("https://ais.sbarc.org/logs_delimited/2018/181231/")) # Returns NULL}
}
\description{
Find siblings, and build urls from a url
}
