% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{.get_links}
\alias{.get_links}
\title{Recursively dive into directories, find urls, fix them, and return the child-most
path}
\usage{
.get_links(u)
}
\description{
Recursively dive into directories, find urls, fix them, and return the child-most
path
}
