% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{.next_xpath}
\alias{.next_xpath}
\title{Build an xpath selector pragmatically from a provided URL}
\usage{
.next_xpath(pat = NULL, ...)
}
\description{
Build an xpath selector pragmatically from a provided URL
}
\examples{
> .next_xpath("https://ais.sbarc.org/logs_delimited/2018/181212/")

[1] ".//a[contains(.,'181212') or contains(@href, '181212')]//following-sibling::a[@href]"
}
