% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{get_ais_data}
\alias{get_ais_data}
\title{Get AIS Data}
\usage{
get_ais_data(links = NULL)
}
\value{
ais_data data.frame
}
\description{
Uses urls2df() function to loop through ais.txt URLs list created by get_ais_urls() function. Creates ais_data data.frame which will write to the database.
}
\details{
Use urls2df() to loop through links to create AIS dataframe
}
\examples{
new_ais_data = update_ais_data(links = links)

}
