% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R, R/utils.R
\name{get_length_km}
\alias{get_length_km}
\title{Get length in km (internal function)}
\usage{
get_length_km(segment)

get_length_km(segment)
}
\arguments{
\item{segment}{(AIS Segment)}
}
\value{
units in km
}
\description{
Get length in km (internal function)

Get kilometer Length from segment
}
\keyword{internal}
