% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R, R/utils.R
\name{get_segment}
\alias{get_segment}
\title{Create segment from 2 points (internal function)}
\usage{
get_segment(p1, p2, crs = 4326)

get_segment(p1, p2, crs = 4326)
}
\arguments{
\item{p1}{point 1}

\item{p2}{point 2}

\item{crs}{(4326)}
}
\value{
linestring

segment
}
\description{
Create segment from 2 points (internal function)

Build segment from AIS Messages
}
\keyword{internal}
