% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{get_segment_data}
\alias{get_segment_data}
\title{Get New AIS Segments Data}
\usage{
get_segment_data(ais_data = NULL)
}
\arguments{
\item{ais_data}{}
}
\value{
segs_data "sf" "data.table" "data.frame"
}
\description{
Get New AIS Segments Data
}
\examples{
new_segs_data = get_segment_data(ais_data = new_ais_data)
}
