% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{get_vsr_ihs_data}
\alias{get_vsr_ihs_data}
\title{Get VSR and IHS data}
\usage{
get_vsr_ihs_data()
}
\value{
vsr_segs_ihs
}
\description{
Get VSR and IHS data
}
\examples{
vsr_ihs_data = get_vsr_ihs_data()
}
