% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.html_doc}
\alias{is.html_doc}
\title{Is an object an xml_document}
\usage{
is.html_doc(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
\code{is.html_doc}
}
\examples{
u <- 'https://ais.sbarc.org/logs_delimited/2019/190202/AIS_SBARC_190202-04.txt'
> is.html_doc(u)
[1] FALSE
> is.html_doc(raw)
[1] TRUE
> is.html_doc(xml_find_all(raw,".//a[1]"))
[1] FALSE

}
