% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.html_nodeset}
\alias{is.html_nodeset}
\title{Determine if an html_nodeset}
\usage{
is.html_nodeset(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
Determine if an html_nodeset
}
\examples{
> is.html_nodeset(xml_find_all(raw,".//a[1]"))
[1] TRUE
> is.html_nodeset(raw)
[1] FALSE
> is.html_nodeset(u)
[1] FALSE

}
