% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.url_only}
\alias{is.url_only}
\title{Determines if an object is a url}
\usage{
is.url_only(x)
}
\arguments{
\item{x}{The object to test}
}
\description{
\code{is.url_only}
}
\examples{
u <- 'https://ais.sbarc.org/logs_delimited/2019/190202/AIS_SBARC_190202-04.txt'
> is.url_only(x = u)
[1] TRUE

raw <- xml2::read_html(u)
> is.url_only(x = raw)
[1] FALSE

> is.url_only(xml_find_all(raw,".//a[1]"))
[1] FALSE

}
