% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{map_ship_limits}
\alias{map_ship_limits}
\title{Interactive map of ship limits}
\usage{
map_ship_limits(lims)
}
\arguments{
\item{lims}{data.frame with sf line segments and speed (km/hr) per segment,
as returned by \code{\link{get_ship_limits}}}
}
\value{
leaflet map of ship segments, color coded by speed (km/hr)
}
\description{
Interactive map of ship limits
}
