% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{map_ship_segments}
\alias{map_ship_segments}
\title{Interactive map of ship segments}
\usage{
map_ship_segments(segs)
}
\arguments{
\item{segs}{data.frame with sf line segments and speed (km/hr) per segment,
as returned by \code{\link[bbnj]{ais_ship_segments}}}
}
\value{
leaflet map of ship segments, color coded by speed (km/hr)
}
\description{
Interactive map of ship segments
}
