% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_stats.R
\name{merge_ihs_vsr}
\alias{merge_ihs_vsr}
\title{Join VSR segments with IHS Ownership Data}
\usage{
merge_ihs_vsr()
}
\arguments{
\item{data}{'vsr_ais_segments' (table from the postgres database)}
}
\value{
vsr_segs_ihs (merged dataframe with VSR segments and ihs data)
}
\description{
Join VSR segments with IHS Ownership Data
}
\examples{
vsr_segs_ihs = merge_ihs_vsr()
}
