% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_stats.R
\name{operator_statistics}
\alias{operator_statistics}
\title{Operator Cooperation Statistics Function ----}
\usage{
operator_statistics(
  data = NULL,
  date_start = NA,
  date_end = NA,
  tonnage = NA,
  ...
)
}
\arguments{
\item{data}{'vsr_ais_segments' (table from the postgres database)}
}
\value{
'operator_stats' (summary statistics for each "operator")
}
\description{
Operator Cooperation Statistics Function ----
}
\examples{

operator_stats = operator_statistics(data=vsr_ihs_data ,date_start = '2018-02-01', date_end = '2019-12-31', tonnage = 300)

Operator statistics from scratch, takes a bit longer
operator_stats_scratch = operator_statistics(date_start = '2018-01-01', date_end = '2019-12-31', tonnage=300)
}
