% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ais_txt}
\alias{read_ais_txt}
\title{Read AIS text file}
\usage{
read_ais_txt(txt)
}
\arguments{
\item{txt}{path to text file}
}
\value{
data.frame of data. TODO: describe all the rules and expectations
  based on AIS_SBARC_*.txt with possibility for alternate file formats and
  filtering for class A position reports, ie Message ID 1,2,3 per
  https://www.navcen.uscg.gov/?pageName=AISMessages. See use of this function in
  https://github.com/mvisalli/shipr/blob/master/data-raw/sbais.R.
}
\description{
Read AIS text file
}
