% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ship_box}
\alias{ship_box}
\title{Show flexdashboard valueBox of ship performance}
\usage{
ship_box(pct_dist_ok, round = 1)
}
\arguments{
\item{pct_dist_ok}{percent distance <= speed limit}
}
\value{
html of valueBox
}
\description{
Show flexdashboard valueBox of ship performance
}
