% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ship_gauge}
\alias{ship_gauge}
\title{Show gauge of ship performance}
\usage{
ship_gauge(pct_dist_ok, round = 1)
}
\arguments{
\item{pct_dist_ok}{percent distance <= speed limit}
}
\value{
htmlwidget of gauge
}
\description{
Show gauge of ship performance
}
