% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{ship_limits}
\alias{ship_limits}
\title{Get breakdown of distance and time above/below speed limit}
\usage{
ship_limits(segs, limit_knots = 10, ship_name, dir_data = "data")
}
\arguments{
\item{segs}{data.frame with sf line segments and speed (km/hr) per segment,
as returned by \code{\link[bbnj]{ais_ship_segments}}}

\item{limit_knots}{speed limit in knots}

\item{ship_name}{used for storing rds data}

\item{dir_data}{directory to cache data}
}
\value{
data.frame summarizing distance and speed, and sf line segments geometry
}
\description{
Get breakdown of distance and time above/below speed limit
}
