% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{ship_segments}
\alias{ship_segments}
\title{Create line segments with speed for specific ship from AIS data}
\usage{
ship_segments(data, ship_name, dir_data = "data")
}
\arguments{
\item{data}{AIS data.frame}

\item{ship_name}{name of ship in AIS data.frame}
}
\value{
data.frame with sf line segments and speed (km/hr) per segment
}
\description{
Create line segments with speed for specific ship from AIS data
}
