% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_stats.R
\name{ship_statistics}
\alias{ship_statistics}
\title{Ship Cooperation Statistics Function ----}
\usage{
ship_statistics(data = NULL, date_start = NA, date_end = NA, tonnage = NA, ...)
}
\arguments{
\item{data}{'vsr_ais_segments' (table from the postgres database)}
}
\value{
'ship_stats' (ship statistics dataframe)
}
\description{
Ship Cooperation Statistics Function ----
}
\examples{

ship_stats = ship_statistics(data=vsr_ihs_data ,date_start = '2018-01-01', date_end = '2019-11-14', tonnage = 300)

ship_statistics from scratch, takes a bit longer
ship_stats = ship_statistics(date_start = '2018-01-01',date_end = '2019-11-14', tonnage=300)
}
