% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{show_ship_box}
\alias{show_ship_box}
\title{Show value box of ship performance}
\usage{
show_ship_box(pct_dist_ok, round = 1)
}
\arguments{
\item{pct_dist_ok}{percent distance <= speed limit}
}
\value{
html of valueBox
}
\description{
Show value box of ship performance
}
