% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{update_ais_data}
\alias{update_ais_data}
\title{Update AIS data and Spatial Features data in postgres database.  TODO: HANDLE ERRORS, WARNINGS, EMPTY TXT FILE URLs}
\usage{
update_ais_data(con, links)
}
\value{
new_ais_data data.frame
}
\description{
Update AIS data and Spatial Features data in postgres database.  TODO: HANDLE ERRORS, WARNINGS, EMPTY TXT FILE URLs
}
\examples{
new_ais_data = update_ais_data(con = con, new_links = new_links)

}
