% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{update_segments_data}
\alias{update_segments_data}
\title{Update AIS Segments Database Table}
\usage{
update_segments_data(segs_data = NULL)
}
\arguments{
\item{ais_data}{}
}
\description{
Creates database connection and writes new ais segments data frame to the "ais_segments" database table.
}
\examples{

}
