% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url.drop_basename}
\alias{url.drop_basename}
\title{Drops the basename}
\usage{
url.drop_basename(url = NULL, ...)
}
\description{
Needed for walking back up the tree
}
\examples{
 u <- "https://ais.sbarc.org/logs_delimited/2019/190202/AIS_SBARC_190202-04.txt"

}
