% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url.hostname}
\alias{url.hostname}
\title{Returns the hostname/protocol portion}
\usage{
url.hostname(url = NULL, ...)
}
\description{
Returns the hostname/protocol portion
}
\examples{
u <- "https://ais.sbarc.org/logs_delimited/2019/190202/AIS_SBARC_190202-04.txt"
> url.hostname(u)
[1] "https://ais.sbarc.org/"

}
