% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{urls2df}
\alias{urls2df}
\title{Read AIS text files into data.frame from URL}
\usage{
urls2df(path = NULL)
}
\arguments{
\item{path}{URL path to AIS text file.}
}
\value{
df data.frame of AIS data.
}
\description{
Read AIS text files into data.frame from URL
}
\details{
based on AIS_SBARC_*.txt with possibility for alternate file formats and filtering for class A position reports, ie Message ID 1,2,3 per https://www.navcen.uscg.gov/?pageName=AISMessages. 
Must provide a url path which leads to an AIS.txt file. AIS data in .txt file must be formatted similarly to those on https://ais.sbarc.org/logs_delimited/
If URL is empty, urls2df() will return a dataframe with 2 observations of 10 variables, but it is essentially to show when there's data gaps from AIS provider.
}
\examples{
 df = urls2df("https://ais.sbarc.org/logs_delimited/2019/190101/AIS_SBARC_190101-00.txt")

Empty file returns small dummy dataframe with date, filename,and system time
df_empty_url = urls2df("https://ais.sbarc.org/logs_delimited/2019/191217/AIS_SBARC_191217-00.txt")

}
