% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{get_segment}
\alias{get_segment}
\title{Create segment from 2 points (internal function)}
\usage{
get_segment(p1, p2, crs = 4326)
}
\arguments{
\item{p1}{sf first point}

\item{p2}{sf second point}

\item{crs}{coordinate reference system (default 4326)}
}
\value{
linestring
}
\description{
Create segment from 2 points (internal function)
}
\keyword{internal}
