% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{update_ais_data}
\alias{update_ais_data}
\title{Update AIS Database Table}
\usage{
update_ais_data(ais_data = NULL)
}
\arguments{
\item{ais_data}{ais_data data.frame}
}
\description{
Update AIS Database Table
}
