% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{update_segments_data}
\alias{update_segments_data}
\title{update_segments_data()}
\usage{
update_segments_data(con, ais_data)
}
\arguments{
\item{ais_data}{}
}
\value{
new_segs_data "sf" "data.table" "data.frame"
}
\description{
update_segments_data()
}
\examples{
new_segs_data = update_segments_data(con = con, ais_data = new_ais_data)
}
