% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ais.R
\name{update_vsr_segments}
\alias{update_vsr_segments}
\title{Update VSR Segments}
\usage{
update_vsr_segments(segs_data = NULL)
}
\arguments{
\item{con}{Formal class PqConnection}
}
\description{
Update VSR Segments
}
\examples{
update_vsr_segments(segs_data = new_segs_data)
}
